% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_highlighted_text.R
\name{get_highlighted_text}
\alias{get_highlighted_text}
\title{Get Highlighted Text}
\usage{
get_highlighted_text()
}
\value{
Character string.
}
\description{
Returns the highlighted text either in the source or the console panes.
}
