% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_console.R
\name{send_console}
\alias{send_console}
\title{Execute code in console}
\usage{
send_console(x)
}
\arguments{
\item{x}{character string. Some code to be executed.}
}
\description{
Wrapper of \code{rstudioapi::sendToConsole}
}
